requirements.txt
# FastAPI
click==8.3.1
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.13.1

# 数据库
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4

# Redis
redis==7.3.0
hiredis==3.3.1

# LLM
langchain==0.3.28
langchain-core==0.3.83
langchain-openai==0.3.35

# OCR
paddleocr==3.4.0
paddlepaddle==3.3.0

# 文档处理
PyMuPDF==1.27.2.2
python-docx==1.2.0
openpyxl==3.1.5
sentence-transformers==5.3.0

# 工具库
python-dotenv==1.2.2
loguru==0.7.3
requests==2.32.5

# 认证
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# 数据处理
pandas==3.0.1
numpy==2.4.1
